BUG: fix error when establishing plugin connections with protocols in all caps
We define the protocol in channel-addresses to be all lowercase. ('pva', 'calc', 'ca', etc)
Used in a while address, it looks like: 'ca://test:example'.

But sometimes we set the protocol to be uppercase, like in the
testing when we use 'CA://MTEST'.

Found this issue when running tests on pyside6, since logger.error() calls are printed and fail
the tests, but on pyqt5 we need to set 'log_cli=true' to get the same
behavior. (will set this in followup patch)